final-project/Bello_system/backend/requirements.txt
Flask==3.1.0
Flask_Cors==5.0.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
PyJWT==2.8.0
serverless-http==3.2.0
google-auth==2.25.2
requests==2.31.0
